OSIS-1489 Case when no object selected before attach + tests
