+++ base/locale/fr_BE/LC_MESSAGES/django.po
@@ -3366,3 +3366,6 @@ msgstr "Liste des formations avec données administratives"
 
 msgid "Validity"
 msgstr "Validité"
+
+msgid "Please Select or Move an item before Attach it"
+msgstr "Veuillez Sélectionner ou Déplacer un élément avant d'Attacher"
