+++ base/views/education_groups/group_element_year/update.py
@@ -36,9 +36,10 @@ from waffle.decorators import waffle_flag
 
 from base.forms.education_group.group_element_year import UpdateGroupElementYearForm
 from base.models.education_group_year import EducationGroupYear
-from base.models.group_element_year import GroupElementYear, get_or_create_group_element_year
+from base.models.group_element_year import GroupElementYear, get_or_create_group_element_year, \
+    get_group_element_year_by_id
 from base.utils.cache import cache
-from base.views.common import display_success_messages
+from base.views.common import display_success_messages, display_warning_messages
 from base.views.common_classes import AjaxTemplateMixin, FlagMixin, RulesRequiredMixin
 from base.views.education_groups import perms
 from base.views.learning_units.perms import PermissionDecoratorWithUser
@@ -48,7 +49,8 @@ from base.views.learning_units.perms import PermissionDecoratorWithUser
 @waffle_flag("education_group_update")
 @PermissionDecoratorWithUser(perms.can_change_education_group, "education_group_year_id", EducationGroupYear)
 def management(request, root_id, education_group_year_id, group_element_year_id):
-    group_element_year = get_object_or_404(GroupElementYear, pk=group_element_year_id)
+    group_element_year_id = int(group_element_year_id)
+    group_element_year = get_group_element_year_by_id(group_element_year_id) if group_element_year_id else None
     action_method = _get_action_method(request)
     source = _get_source(request)
     response = action_method(
@@ -93,15 +95,20 @@ def _detach(request, group_element_year, *args, **kwargs):
 
 @require_http_methods(['GET', 'POST'])
 def _attach(request, group_element_year, *args, **kwargs):
-    new_parent = EducationGroupYear.objects.get(id=kwargs.get('education_group_year_id'))
-    success_msg = _("Attached to %(acronym)s") % {'acronym': new_parent}
-    child_id = int(cache.get('child_to_cache_id'))
-    get_or_create_group_element_year(
-        parent=new_parent,
-        child=EducationGroupYear.objects.get(id=child_id)
-    )
-    cache.set('education_group_year_id', None, timeout=None)
-    display_success_messages(request, success_msg)
+    child_to_cache_id = cache.get('child_to_cache_id')
+    if child_to_cache_id:
+        new_parent = EducationGroupYear.objects.get(id=kwargs.get('education_group_year_id'))
+        success_msg = _("Attached to %(acronym)s") % {'acronym': new_parent}
+        child_id = int(child_to_cache_id)
+        get_or_create_group_element_year(
+            parent=new_parent,
+            child=EducationGroupYear.objects.get(id=child_id)
+        )
+        cache.set('child_to_cache_id', None, timeout=None)
+        display_success_messages(request, success_msg)
+    else:
+        warning_msg = _("Please Select or Move an item before Attach it")
+        display_warning_messages(request, warning_msg)
 
 
 def _get_action_method(request):
